+++ src/countScenariosByStatuses.spec.ts
@@ -1,17 +1,21 @@
 import { SupportCode } from '@cucumber/fake-cucumber'
 import { Query as GherkinQuery } from '@cucumber/gherkin-utils'
 import * as messages from '@cucumber/messages'
-import { SourceReference, TestStepResultStatus } from '@cucumber/messages'
+import { Envelope, SourceReference, TestStepResultStatus } from '@cucumber/messages'
 import { Query as CucumberQuery } from '@cucumber/query'
+import fs from 'fs'
+import path from 'path'
 
 import { runFeature } from '../test-utils'
 import countScenariosByStatuses from './countScenariosByStatuses'
+import { EnvelopesQuery } from './EnvelopesQueryContext'
 
 const sourceReference: SourceReference = {}
 
 describe('countScenariosByStatuses', () => {
   let gherkinQuery: GherkinQuery
   let cucumberQuery: CucumberQuery
+  let envelopesQuery: EnvelopesQuery
   let supportCode: SupportCode
 
   jest.setTimeout(3000)
@@ -19,6 +23,7 @@ describe('countScenariosByStatuses', () => {
   beforeEach(() => {
     gherkinQuery = new GherkinQuery()
     cucumberQuery = new CucumberQuery()
+    envelopesQuery = new EnvelopesQuery()
     supportCode = new SupportCode()
     supportCode.defineStepDefinition(sourceReference, 'a passed step', () => null)
     supportCode.defineStepDefinition(sourceReference, 'a failed step', () => {
@@ -45,10 +50,11 @@ Feature: statuses
     const envelopes = await runFeature(feature, gherkinQuery, supportCode)
     for (const envelope of envelopes) {
       cucumberQuery.update(envelope)
+      envelopesQuery.update(envelope)
     }
 
     const { scenarioCountByStatus, statusesWithScenarios, totalScenarioCount } =
-      countScenariosByStatuses(gherkinQuery, cucumberQuery)
+      countScenariosByStatuses(gherkinQuery, cucumberQuery, envelopesQuery)
 
     expect(scenarioCountByStatus[messages.TestStepResultStatus.PASSED]).toEqual(2)
     expect(scenarioCountByStatus[messages.TestStepResultStatus.FAILED]).toEqual(1)
@@ -79,10 +85,11 @@ Feature: statuses
     const envelopes = await runFeature(feature, gherkinQuery, supportCode)
     for (const envelope of envelopes) {
       cucumberQuery.update(envelope)
+      envelopesQuery.update(envelope)
     }
 
     const { scenarioCountByStatus, statusesWithScenarios, totalScenarioCount } =
-      countScenariosByStatuses(gherkinQuery, cucumberQuery)
+      countScenariosByStatuses(gherkinQuery, cucumberQuery, envelopesQuery)
 
     expect(scenarioCountByStatus[messages.TestStepResultStatus.PASSED]).toEqual(1)
     expect(scenarioCountByStatus[messages.TestStepResultStatus.FAILED]).toEqual(1)
@@ -93,5 +100,30 @@ Feature: statuses
       TestStepResultStatus.UNDEFINED,
     ])
     expect(totalScenarioCount).toEqual(3)
+  })
+
+  it('only includes pickles that were slated for execution as test cases', () => {
+    const raw = fs.readFileSync(
+      path.join(__dirname, '../test-utils/messages/filtered-pickles.ndjson'),
+      {
+        encoding: 'utf-8',
+      }
+    )
+    const envelopes: Envelope[] = JSON.parse('[' + raw.trim().split('\n').join(',') + ']')
+    const gherkinQuery = new GherkinQuery()
+    const cucumberQuery = new CucumberQuery()
+    envelopes.forEach((envelope) => {
+      gherkinQuery.update(envelope)
+      cucumberQuery.update(envelope)
+      envelopesQuery.update(envelope)
+    })
+
+    const { totalScenarioCount } = countScenariosByStatuses(
+      gherkinQuery,
+      cucumberQuery,
+      envelopesQuery
+    )
+
+    expect(totalScenarioCount).toEqual(1)
   })
 })